Complete exercises up to 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -20,12 +20,12 @@ blogsRouter.get("/:id", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
+
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token invalid" });
   }
   const user = await User.findById(decodedToken.id);
-
   if (!user) {
     return response.status(400).json({ error: "userId missing or not valid" });
   }
@@ -46,6 +46,22 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  if (!user) {
+    return response.status(400).json({ error: "userId missing or not valid" });
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response
+      .status(401)
+      .json({ error: "you are not authorized to delete this blog" });
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
